Fix allenai/nora-issues-research#187 (mcp freeze on long output)
As I was finishing this up, I realized that MCP seems to allow returning a list
of `ContextText` instead of one big string.  It's possible that this would fix
the size limit (ie. chunk into a list that we return via MCP instead of sending
multiple responses), though I haven't confirmed if it actually works.