Disable DV cache to ensure valid token counts
If requests are cached, then the token usage will show up as 0.  It doesn't
seem to matter whether we use `total_usage()` or `actual_usage()` on the cached
client.  We could potentially work around it if we instead patched `create()`
and `create_stream()` methods to inspect the cached usage values or directly
count tokens in the messages, but for now it's much simpler to just disable the
cache.